SF-210 - Table now initialises with correct row heights for word-wrapped rows
If word wrapping is enabled, then UpdateScrollBars() is called just
after the first paint event, so that scrollbars are correctly
calculated, taking into account the actual rendered height of rows.
Without this you sometimes did not see the scrollbar when XPTable
loaded, as it had assumed the default row height for all rows when
deciding to show the scrollbar of not.