Reworked SendMessage and PostMessage pinvokes
Fix for issue #38 - The SendMessage and PostMessage pinvokes don't have
the proper signatures. `wParam` and `lParam` should be declared as
IntPtrs since their size is 32-bit/64-bit depending on the platform.
SendMessage's `LRESULT` return type should also be modeled as such.
These bad declarations can cause the stack to become unbalanced, causing
a crash. However, it seems that these were still working by the skin of
their teeth due to how these values are handled under 64-bit calling
conventions - they're passed in registers in 64-bit, and passed on the
stack (iirc) in 32-bit. Because the values were always 32-bit but they
they're not passed on the stack in 64-bit, there's no stack to
unbalance, so it didn't crash.